@@ -14,7 +14,7 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.9.0 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
   bloc_concurrency: ^0.2.0
