@@ -10,7 +10,7 @@ abstract class LoginEvent extends Equatable {
 class EmailChanged extends LoginEvent {
   final String email;
 
-  const EmailChanged({@required this.email});
+  const EmailChanged({required this.email});
 
   @override
   List<Object> get props => [email];
@@ -22,7 +22,7 @@ class EmailChanged extends LoginEvent {
 class PasswordChanged extends LoginEvent {
   final String password;
 
-  const PasswordChanged({@required this.password});
+  const PasswordChanged({required this.password});
 
   @override
   List<Object> get props => [password];
@@ -36,8 +36,8 @@ class Submitted extends LoginEvent {
   final String password;
 
   const Submitted({
-    @required this.email,
-    @required this.password,
+    required this.email,
+    required this.password,
   });
 
   @override
@@ -54,7 +54,7 @@ class LoginWithGooglePressed extends LoginEvent {}
 class ForgotPasswordPressed extends LoginEvent {
   final String email;
 
-  const ForgotPasswordPressed({@required this.email});
+  const ForgotPasswordPressed({required this.email});
 
   @override
   List<Object> get props => [email];
@@ -68,8 +68,8 @@ class LoginWithCredentialsPressed extends LoginEvent {
   final String password;
 
   const LoginWithCredentialsPressed({
-    @required this.email,
-    @required this.password,
+    required this.email,
+    required this.password,
   });
 
   @override
