@@ -12,13 +12,13 @@ class LoginState {
   bool get isFormValid => isEmailValid && isPasswordValid;
 
   LoginState(
-      {@required this.isEmailValid,
-      @required this.isPasswordValid,
-      @required this.isSubmitting,
-      @required this.isSuccess,
-      @required this.isFailure,
-      @required this.isPasswordResetMailSent,
-      @required this.isPasswordResetFailure});
+      {required this.isEmailValid,
+      required this.isPasswordValid,
+      required this.isSubmitting,
+      required this.isSuccess,
+      required this.isFailure,
+      required this.isPasswordResetMailSent,
+      required this.isPasswordResetFailure});
 
   factory LoginState.empty() {
     return LoginState(
@@ -92,8 +92,8 @@ class LoginState {
   }
 
   LoginState update({
-    bool isEmailValid,
-    bool isPasswordValid,
+    bool? isEmailValid,
+    bool? isPasswordValid,
   }) {
     return copyWith(
       isEmailValid: isEmailValid,
@@ -106,15 +106,16 @@ class LoginState {
     );
   }
 
-  LoginState copyWith(
-      {bool isEmailValid,
-      bool isPasswordValid,
-      bool isSubmitEnabled,
-      bool isSubmitting,
-      bool isSuccess,
-      bool isFailure,
-      bool isPasswordResetFailure,
-      bool isPasswordResetMailSent}) {
+  LoginState copyWith({
+    bool? isEmailValid,
+    bool? isPasswordValid,
+    bool? isSubmitEnabled,
+    bool? isSubmitting,
+    bool? isSuccess,
+    bool? isFailure,
+    bool? isPasswordResetFailure,
+    bool? isPasswordResetMailSent,
+  }) {
     return LoginState(
       isEmailValid: isEmailValid ?? this.isEmailValid,
       isPasswordValid: isPasswordValid ?? this.isPasswordValid,
