@@ -1,6 +1,6 @@
 import 'package:shared_preferences/shared_preferences.dart';
 
 class Prefer {
-  static SharedPreferences prefs;
+  static late SharedPreferences prefs;
   static int themeIndexPref = 0;
 }
