@@ -1,11 +1,9 @@
 import 'package:flutter/material.dart';
 
 class RegisterButton extends StatelessWidget {
-  final VoidCallback _onPressed;
+  final VoidCallback? onPressed;
 
-  const RegisterButton({Key key, VoidCallback onPressed})
-      : _onPressed = onPressed,
-        super(key: key);
+  const RegisterButton({Key? key, this.onPressed}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -13,7 +11,7 @@ class RegisterButton extends StatelessWidget {
       width: 70.0,
       height: 30.0,
       child: TextButton(
-        onPressed: _onPressed,
+        onPressed: onPressed,
         style: TextButton.styleFrom(
           backgroundColor: Color(0xFF5d74e3),
           disabledBackgroundColor: Colors.grey,
