@@ -7,7 +7,7 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 class RegisterScreen extends StatelessWidget {
   final UserRepository _userRepository;
 
-  RegisterScreen({Key key, @required UserRepository userRepository})
+  RegisterScreen({Key? key, required UserRepository userRepository})
       : assert(userRepository != null),
         _userRepository = userRepository,
         super(key: key);
