@@ -10,18 +10,21 @@ class EditTodoForm extends StatefulWidget {
   final OnSaveCallback onSave;
   final Todo todo;
 
-  EditTodoForm(
-      {Key key, @required this.isEditing, @required this.onSave, this.todo})
-      : super(key: key);
+  EditTodoForm({
+    Key? key,
+    required this.isEditing,
+    required this.onSave,
+    required this.todo,
+  }) : super(key: key);
 
   @override
   _EditTodoFormState createState() => _EditTodoFormState();
 }
 
 class _EditTodoFormState extends State<EditTodoForm> {
-  String _task;
-  String _note;
-  DateTime _dateTime;
+  late String _task;
+  late String _note;
+  DateTime? _dateTime;
 
   bool get isEditing => widget.isEditing;
 
@@ -91,22 +94,22 @@ class _EditTodoFormState extends State<EditTodoForm> {
                                 padding: EdgeInsets.only(right: 22.0),
                                 child: TextButton(
                                   onPressed: () {
-                                    if (_formKey.currentState.validate()) {
-                                      _formKey.currentState.save();
+                                    if (_formKey.currentState!.validate()) {
+                                      _formKey.currentState!.save();
                                       widget.onSave(
-                                          _task,
-                                          _note,
-                                          _dateTime == null
-                                              ? 0
-                                              : _dateTime
-                                                  .millisecondsSinceEpoch);
+                                        _task,
+                                        _note,
+                                        _dateTime == null
+                                            ? 0
+                                            : _dateTime!.millisecondsSinceEpoch,
+                                      );
                                       Navigator.pop(context);
                                     }
                                   },
                                   style: TextButton.styleFrom(
                                     backgroundColor: Theme.of(context)
                                         .buttonTheme
-                                        .colorScheme
+                                        .colorScheme!
                                         .background,
                                     disabledBackgroundColor: Colors.grey,
                                     shape: RoundedRectangleBorder(
@@ -172,11 +175,11 @@ class _EditTodoFormState extends State<EditTodoForm> {
                                       isEditing ? widget.todo.task : '',
                                   autofocus: isEditing,
                                   validator: (val) {
-                                    return val.trim().isEmpty
+                                    return val!.trim().isEmpty
                                         ? 'Please enter some text'
                                         : null;
                                   },
-                                  onSaved: (value) => _task = value,
+                                  onSaved: (value) => _task = value!,
                                   cursorColor: Color(0xFF5d74e3),
                                   style: Theme.of(context).textTheme.headline1,
                                   decoration: InputDecoration(
@@ -219,10 +222,10 @@ class _EditTodoFormState extends State<EditTodoForm> {
                                     initialValue:
                                         isEditing ? widget.todo.note : '',
                                     maxLines: null,
-                                    onSaved: (value) => _note = value,
+                                    onSaved: (value) => _note = value!,
                                     autofocus: !isEditing,
                                     validator: (val) {
-                                      return val.trim().isEmpty
+                                      return val!.trim().isEmpty
                                           ? 'Please enter some text'
                                           : null;
                                     },
@@ -293,7 +296,7 @@ class _EditTodoFormState extends State<EditTodoForm> {
                                             )
                                           : Text(
                                               DateFormat('EEE d MMM hh:mm a')
-                                                  .format(_dateTime.toLocal()),
+                                                  .format(_dateTime!.toLocal()),
                                               style: Theme.of(context)
                                                   .textTheme
                                                   .headline1,
