@@ -6,12 +6,12 @@ typedef OnSaveCallback = Function(String task, String note);
 class AddEditScreen extends StatefulWidget {
   final bool isEditing;
   final OnSaveCallback onSave;
-  final Todo todo;
+  final Todo? todo;
 
   AddEditScreen({
-    Key key,
-    @required this.onSave,
-    @required this.isEditing,
+    Key? key,
+    required this.onSave,
+    required this.isEditing,
     this.todo,
   }) : super(key: key);
 
@@ -22,8 +22,8 @@ class AddEditScreen extends StatefulWidget {
 class _AddEditScreenState extends State<AddEditScreen> {
   static final GlobalKey<FormState> _formKey = GlobalKey<FormState>();
 
-  String _task;
-  String _note;
+  late String _task;
+  late String _note;
 
   bool get isEditing => widget.isEditing;
 
@@ -44,25 +44,25 @@ class _AddEditScreenState extends State<AddEditScreen> {
           child: ListView(
             children: [
               TextFormField(
-                initialValue: isEditing ? widget.todo.task : '',
+                initialValue: isEditing ? widget.todo!.task : '',
                 autofocus: !isEditing,
                 style: textTheme.headline5,
                 decoration: InputDecoration(
                   hintText: 'What needs to be done?',
                 ),
                 validator: (val) {
-                  return val.trim().isEmpty ? 'Please enter some text' : null;
+                  return val!.trim().isEmpty ? 'Please enter some text' : null;
                 },
-                onSaved: (value) => _task = value,
+                onSaved: (value) => _task = value!,
               ),
               TextFormField(
-                initialValue: isEditing ? widget.todo.note : '',
+                initialValue: isEditing ? widget.todo!.note : '',
                 maxLines: 10,
                 style: textTheme.subtitle1,
                 decoration: InputDecoration(
                   hintText: 'Additional Notes...',
                 ),
-                onSaved: (value) => _note = value,
+                onSaved: (value) => _note = value!,
               )
             ],
           ),
@@ -72,8 +72,8 @@ class _AddEditScreenState extends State<AddEditScreen> {
         tooltip: isEditing ? 'Save changes' : 'Add Todo',
         child: Icon(isEditing ? Icons.check : Icons.add),
         onPressed: () {
-          if (_formKey.currentState.validate()) {
-            _formKey.currentState.save();
+          if (_formKey.currentState!.validate()) {
+            _formKey.currentState!.save();
             widget.onSave(_task, _note);
             Navigator.pop(context);
           }
