@@ -2,19 +2,19 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:equatable/equatable.dart';
 
 class TodoEntity extends Equatable {
-  final bool complete;
+  final bool? complete;
   final String id;
   final String note;
   final String task;
-  final String userId;
-  final int time;
+  final String? userId;
+  final int? time;
 
   const TodoEntity(
       this.task, this.id, this.note, this.complete, this.userId, this.time);
 
-  Map<String, Object> toJson() {
+  Map<String, Object?> toJson() {
     return {
-      "complete": complete,
+      "complete": complete ?? false,
       "task": task,
       "note": note,
       "id": id,
@@ -24,7 +24,7 @@ class TodoEntity extends Equatable {
   }
 
   @override
-  List<Object> get props => [complete, id, note, task, userId, time];
+  List<Object> get props => [id, note, task];
 
   @override
   String toString() {
@@ -42,12 +42,12 @@ class TodoEntity extends Equatable {
   }
 
   static TodoEntity fromSnapshot(DocumentSnapshot snap) {
-    Map<String, dynamic> data = snap.data();
+    dynamic data = snap.data();
     return TodoEntity(data['task'], snap.id, data['note'], data['complete'],
         data['userId'], data['time']);
   }
 
-  Map<String, Object> toDocument() {
+  Map<String, Object?> toDocument() {
     return {
       "complete": complete,
       "task": task,
