@@ -3,9 +3,10 @@ import 'package:flutter/material.dart';
 class SocialIcon extends StatelessWidget {
   final List<Color> colors;
   final IconData iconData;
-  final Function onPressed;
+  final void Function()? onPressed;
 
-  SocialIcon({this.colors, this.iconData, this.onPressed});
+  SocialIcon(
+      {required this.colors, required this.iconData, required this.onPressed});
 
   @override
   Widget build(BuildContext context) {
